feat: limits post responses to 201 success code
We expect our POST endpoints to be creation endpoints, and so the response should be a 201 and supply an ID (which is caught by another rule).
This rule limits the spec to being _only_ 201, as it seems possible to allow many 2xx responses.
Outside of the 2xx range I've been pretty open with the ranges being 3xx, 4xx and 5xx.