fix: npm build and semantic-release
Ok, this seems to fix issues with sweater-comb not being
npm-installable.

Queue: "Fantasic Voyage"...

So what I found when `yarn pack`ing this package was:
- It can be `yarn add`ed to yarn projects just fine and sweater-comb /
  optic works.
- It cannot be `npm install`ed into npm projects. The ajv module was
  failing to initialize, which is what the past few 1.0.1-present
  releases were trying to fix.

The root cause seems to be that npm seems to pick the wrong ajv version,
but yarn gets it right. ajv is a transitive dependency, pulled in via
the spectral rules support at the very least.

This tool needs to be used in an npm-based project, so "just use yarn"
wouldn't be an acceptable limitation. I found a workaround:

- Bundling all the Spectral and Optic dependencies seems to satisfy
  `npm install`, and the sweater-comb script works!
- `yarn pack` can't deal with bundling dependencies, it bombs out in the
  presence of them, and the project seems to declare it a won't fix
  situation.

Queue: "Stuck in the middle with you"...

So what this does is:
- In `package.json`, bundle the dependencies that npm will otherwise break.
- Build this project with yarn.
- Pack it with `npm pack`.
- Release it with npx semantic-release, which is npm-friendly.

I've tested sweater-comb works when `npm install`/`yarn add` -ed to
projects managed with each.

Other changes in this PR are undo-ing the cargo-culting I was doing in
trying to appease the node module loader.