chore: use effectiveOn for select rules
Make the tenancy and operationId naming rules effective since
2021-07-01. This allows some already-released early v3
implementations to pass CI.

This also changes the frame of reference to the version date of the
resource, rather than the date on which the change is occurring. This
allows CI to pass on non-breaking changes to releases made prior to
later rules.

Side note: This does raise the need for a rule which prevents
back-dating a new release past a certain point. There are several
reasons back-dating releases is a problem (besides just circumventing
rules) -- it can play havoc with deprecation & sunsetting, as well as
surprise users.