fix: JSON API rules are requirementOnChange
Making JSON API rules an invariant requirement gets us into a catch-22
situation (invariant meaning, the rule is applied to all specs as they
are at a point in time, rather than on a changeset) where we have
existing APIs out of compliance, but we can't turn on the rule because
they will never pass.

This changes the JSON API rules to "requirementOnChange", so that the
rule only applies to proposed changes in CI.

Also adding the `schemas` directory to the package build, which was
causing the rule to error out in our built NPM and Docker packages.