fix: fetch OpenAPI specs and compare locally with optic-ci CLI
Replace Optic TestHelpers and loadSpecFromUrl with HTTP GET and shelling
out to execute optic-ci.

Several reasons for doing this:
- loadSpecFromUrl did not provide enough control in handling HTTP responses.
- Running the check with the optic-ci CLI allows Optic Cloud integration
  and reduces the potential for drift between development and
  production.
- Shelling out should reduce the memory footprint, as the main resident
  process handling readiness probes will be small. It does not load
  Sweater Comb or Optic modules directly.

This raises a question, why not write this sidecar in Go? We could, but
it would then need to be released lock-step with Sweater Comb. Combining
it into the same NPM package and OCI distribution reduces configuration
and potential for drift. It is more important to govern with the right
rules.