fix: regression in operation naming rules for experimental stability
Some of the operation rules ported over to the new DSL were matching
!isBreakingChangeAllowed where they shouldn't have.

The result was a small window where operation ID and parameter naming
rules were not applied to experimental.

Technically this change introduces a tighter constraint which would
normally require a new REST API rules version.

However since this is a regression, which affected only experimental
(which is allowed breaking changes) and the regression was only released
for a small window of several days, the potential impact to fixing this
in-place should be small. Projects will need to conform to these rules
in order to promote to beta anyway.

In the event this change does disrupt API work, affected projects may
pin their Sweater Comb dependency version to buy themselves some time to
update their APIs.

Drive-by: increase timeout for a test flapping in Circle CI