fix: do not allow parameters in the path root
Do not allow resource paths to declare a parameter in the path root.

This is currently a _de facto_ rule, because API gateway administrators
will not allow catch-all paths at the root like this. This PR makes it
_de jure_.

It's a new constraint, but no standards bump required, because it would
never have been routed in the first place.

Fixes #253