fix: regression in batch POST rule
This fixes a regression in the batch POST rule which allows a 204
response when the request body `data` property is an array.

When the rule was ported over, it seems to have changed to expect the
schema of the request body to be an array. It should instead be of the
form:

```
{
  data: [{id: "some-id", type: "some-type", ...}]
}
```