fix: allow breaking removals in a spec being removed
When a spec is sunset it can be removed; the lifecycle checks are done
on the compiled version to ensure that the spec can be removed, but
sweater-comb complains about the removal of properties and operations as
part of the spec being deleted.

Since there's a check to ensure it's safe for the spec to be removed,
this commit allows the removal of operations and properties as part of
the spec deletion.