fix: bulk compare paths relative to git repo, async catch
When using bulk-compare with glob, paths must be relative to the git
repository. This may not be the same as the .vervet.yaml directory in
cases where the API project directory is a subdirectory in the cloned
git repo.

Improve exception handling to allow all asynchronous execution to
complete in the top-level catch. It turns out that exiting the process
in the catch may terminate the process before some logged output has a
chance to be written to the terminal output, resulting in a non-zero
exit code with no apparent reason to the user.