fix: allow resource IDs to be format: uri
Some resources are uniquely located by a structured identifier, and do
not have a UUID. Examples include packages (identified by a pURL),
issues, and CVEs.

For IDs which are not UUIDs, allow the use of a URI to identify the
resource. Even if the URI is of a custom type, requiring a URI creates
the right incentive to define a well-formed structure for that
identifier.