docs: adds timestamp filters to rest standards
The REST standards provide rules for timestamp properties. A natural use
of timestamp properties is filtering for resources between two dates.
This documents a pattern already in use for filter values using _before
and _after suffixes along with the existing _at suffix.

As a drive by, this branch also updates references to "v3" to "REST".