feat: adds linting for a POST to a relationship
According to the JSON:API spec, `POST`s to a relationship of a resource do not need the `attributes` object in the `POST` body, which we require for all other `POST` requests.
This allows the shape of the body to be different in the case of relationships by looking at the path to determine whether it is a relationship. As `relationship` is a reserved word, then this should be all that is needed to cover it.