fix: remove attributes from required post body properties
The upstream JSON:API spec states that only type is required. Requiring
attributes means we will be unable to create resources that have no
attributes.

An example of what this is blocking is creating app_install resources
which have bodies similar to:

    {
        "type": "app_install",
        "relationships": {
            "app": { "data": { "type": "app", "id": 1234 } }
        }
    }

This is perfectly valid under JSON:API specs but fails the current
linting due to the missing attributes.