fix: ship built version of package
When we made the switch from yarn to npm, the command to build the
project was accidentally removed. As such we are currently just shipping
the sources and not the built binary, so any project that consumes this
will have to build it themselves.

This re-adds the build step so sweater-comb can be ran using npx again.