chore: downgrade optic 0.43.0
Fixes #472

I'm not sure why this fixes the problem, but I bisected and found it
started with 2.1.1.

That release included an upgrade of the @useoptic/optic dependency.
Reverting that upgrade seemed to fix the problem.

There's likely an Optic bug here to follow up on, but this unblocks us
in the meantime.