feat: add nix flake
This allows easy development with the devshell containing what you need
to work on the project, as well as a package for consuming this package
independent of npm.

Can be consumed with:
nix run github:snyk/sweater-comb

Using date based versioning as we don't have good tooling to update the
version here on a release and we can't expect everyone who touches the
project to care about nix.