fix: exit cleanly on informational commands
When intercepting exit codes from commander, it throws an exception when
version or help commands are requested. As we were not catching these
were were exiting with code 1, which is surprising.

We need to intercept the exit code as we have some async commands and
commander may exit early on error before logs are printed.