fix: add nix result to path
This package builds @snyk/sweater-comb, which means the sweater-comb
binary is not on the exported path from the resulting nix derivation.
This is fine when running the result directly as it is picked up from
mainProgram - however when consuming into a devShell it means that the
binary is not in the users path.