fix: move header examples under schema
Rapidoc expects the example to be under the `schema` key, as such we are
not currently rendering any example - except for VersionLifecycleStage
which for some reason is correct.

This will not immediately fix all the endpoints, teams will need to
update their references to point to the latest version of this doc.