fix: assume parsed documents are OpenAPI v3
Optic does support OpenAPI v2, but we do not. The recent update to
optic/openapi-io adds type parameters for OpenAPI v2 which are
incompatible with our assumptions that all specs are v3. This just adds
type annotations for our assumptions, it will break if we ever try to
parse a v2 document.