fix: update Optic to 1.0.0 to prevent cloud auth
Previous versions of Optic tried to auth to their cloud service if a api
key was set in the env. Recently Optic have shut down their cloud
service so this call is failing and causing the whole process to error
out. We have a valid API key set in the CircleCI context that most
pipelines run because at one point we were planning on using the
service.