feat: allow namespaced component names
OpenAPI generated by Typespec will use namespace-qualified component names in
the generated OpenAPI, when types are in a different namespace from that of the
OpenAPI spec. Additionally, components derived from model properties such as
parameters or headers use the property name as a suffix.

This updates the componentNameCase rule to allow PascalCase component names
with an optional dot.case namespace prefix and optional snake_case suffix.

Updated the sweater-comb standards documentation with this change.

DGP-50